projects/KnowledgeBuilder/requirements.txt
python-dotenv==1.2.1
pydantic==2.12.5
requests==2.31.0
Flask==3.0.0
anthropic==0.75.0
openai==2.14.0
lancedb==0.26.1
numpy==2.4.0
pandas==2.3.3
click==8.3.1
rich==14.2.0
loguru==0.7.3
pyyaml==6.0.3
httpx==0.28.1
